Fixed process destructor bug, simple test works.
For some reason processes were getting destroyed that were never constructed
(i.e., they were {0} in memory).  I don't know why, but this was causing
stdin to get closed.

Likely guess is that emplace_back isn't doing what it's supposed to, and is
performing a move construction into the vector, but these elements were never
constructed, so the swap() swaps an invalid element into the reference.

This is most likely a bug in stdlib.  I'll look into it.

TODO:  The help message, maybe even write a manpage...